--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "claude-playwright",
-  "version": "0.1.0-alpha.21",
+  "version": "0.1.0-alpha.22",
   "description": "Seamless integration between Claude Code and Playwright MCP for efficient browser automation and testing",
   "main": "dist/index.cjs",
   "types": "dist/index.d.ts",