… src/core/feature-flag-manager.ts
… src/core/performance-monitor.ts
… src/core/test-orchestrator.ts
… src/core/validation-reporter.ts
… tests/unit/cache/bidirectional-cache.test.ts
… tests/unit/cache/dom-signature.test.ts
… tests/unit/circuit-breaker/state-management.test.ts